catalog/requirements.txt
Flask
mysql-connector
flask-mysql
flask-cors